+++ packages/docs/content/.vitepress/config.js
@@ -1,7 +1,6 @@
 /**
  * @type {import('vitepress').UserConfig<import('vitepress').DefaultTheme.Config>}
  */
-
 module.exports = {
   lang: 'en-US',
   markdown: {
@@ -12,11 +11,14 @@ module.exports = {
   title: 'PDF Donkey',
   description: 'Generate PDFs on the fly.',
   themeConfig: {
+    nextLinks: true,
+    prevLinks: true,
     repo: 'soulsam480/pdf-donkey',
-    docsDir: 'docs',
+    docsDir: 'packages/docs',
     editLinks: true,
     editLinkText: 'Edit this page on GitHub',
     lastUpdated: 'Last Updated',
+    logo: '/donkey-trans.png',
     nav: [
       { text: 'Guide', link: '/guide/index.html', activeMatch: '^/guide/' },
       {
@@ -24,5 +26,21 @@ module.exports = {
         link: 'https://github.com/soulsam480/pdf-donkey/releases',
       },
     ],
+    sidebar: {
+      '/guide/': getSidebar(),
+      '/': getSidebar(),
+    },
   },
 };
+
+function getSidebar() {
+  return [
+    {
+      text: 'Introduction',
+      children: [
+        { text: 'Getting started', link: '/guide/' },
+        { text: 'Create new template', link: '/guide/create-new-template' },
+      ],
+    },
+  ];
+}